+++ lbrynet/lbrynet_daemon/Daemon.py
@@ -105,6 +105,7 @@ class FileID:
     NAME = 'name'
     SD_HASH = 'sd_hash'
     FILE_NAME = 'file_name'
+    STREAM_HASH = 'stream_hash'
 
 
 # TODO add login credentials in a conf file
@@ -819,17 +820,13 @@ class Daemon(AuthJSONRPCServer):
         """
         timeout = timeout if timeout is not None else conf.settings['download_timeout']
 
-        try:
-            helper = _DownloadNameHelper(self, name, timeout, download_directory, file_name,
-                                         wait_for_write)
-        except Exception as err:
-            log.exception(err)
-            raise err
 
+        helper = _DownloadNameHelper(self, name, timeout, download_directory, file_name,
+                                         wait_for_write)
         if not stream_info:
             self.waiting_on[name] = True
             stream_info = yield self._resolve_name(name)
-            del self.waiting_on[name]
+
         lbry_file = yield helper.setup_stream(stream_info)
         sd_hash, file_path = yield helper.wait_or_get_stream(stream_info, lbry_file)
         defer.returnValue((sd_hash, file_path))
@@ -1029,6 +1026,12 @@ class Daemon(AuthJSONRPCServer):
                 return lbry_file
         raise Exception("File %s not found" % file_name)
 
+    def _find_lbry_file_by_stream_hash(self, stream_hash):
+        for lbry_file in self.lbry_file_manager.lbry_files:
+            if lbry_file.stream_hash == stream_hash:
+                return lbry_file
+        raise NoSuchStreamHash(stream_hash)
+
     @defer.inlineCallbacks
     def _get_lbry_file_by_uri(self, name):
         try:
@@ -1050,6 +1053,11 @@ class Daemon(AuthJSONRPCServer):
         defer.returnValue(lbry_file)
 
     @defer.inlineCallbacks
+    def _get_lbry_file_by_stream_hash(self, stream_hash):
+        lbry_file = yield self._find_lbry_file_by_stream_hash(stream_hash)
+        defer.returnValue(lbry_file)
+
+    @defer.inlineCallbacks
     def _get_lbry_file(self, search_by, val, return_json=True):
         helper = _GetFileHelper(self, search_by, val, return_json)
         try:
@@ -1057,7 +1065,7 @@ class Daemon(AuthJSONRPCServer):
             defer.returnValue(lbry_file)
         except Exception as err:
             # TODO: do something with the error, don't return None when a file isn't found
-            defer.returnValue(None)
+            defer.returnValue(False)
 
     def _get_lbry_files(self):
         def safe_get(sd_hash):
@@ -1376,7 +1384,10 @@ class Daemon(AuthJSONRPCServer):
         Returns:
             list
         """
-        return self._render_response(sorted(self.callable_methods.keys()))
+        return self._render_response(sorted(
+            [command for command in self.callable_methods.keys()
+             if 'DEPRECATED' not in getattr(self, "jsonrpc_" + command).__doc__]
+        ))
 
     def jsonrpc_get_balance(self):
         """
@@ -1453,22 +1464,35 @@ class Daemon(AuthJSONRPCServer):
 
     def jsonrpc_file_get(self, **kwargs):
         """
-        Get a file
+        Get a file, if no matching file exists returns False
 
         Args:
             'name': get file by lbry uri,
             'sd_hash': get file by the hash in the name claim,
             'file_name': get file by its name in the downloads folder,
+            'stream_hash': get file by its stream hash
         Returns:
-            'completed': bool
-            'file_name': string
-            'key': hex string
-            'points_paid': float
-            'stopped': bool
-            'stream_hash': base 58 string
-            'stream_name': string
-            'suggested_file_name': string
-            'sd_hash': string
+            'completed': bool,
+            'file_name': str,
+            'download_directory': str,
+            'points_paid': float,
+            'stopped': bool,
+            'stream_hash': str (hex),
+            'stream_name': str,
+            'suggested_file_name': str,
+            'sd_hash': str (hex),
+            'lbry_uri': str,
+            'txid': str (b58),
+            'claim_id': str (b58),
+            'download_path': str,
+            'mime_type': str,
+            'key': str (hex),
+            'total_bytes': int,
+            'written_bytes': int,
+            'code': str,
+            'message': str
+            'metadata': Metadata dict if claim is valid, otherwise status str
+        }
         """
         d = self._get_deferred_for_lbry_file(kwargs)
         d.addCallback(lambda r: self._render_response(r))
@@ -1557,6 +1581,7 @@ class Daemon(AuthJSONRPCServer):
             'stream_hash': hex string
             'path': path of download
         """
+
         timeout = timeout if timeout is not None else self.download_timeout
         download_directory = download_directory or self.download_directory
         sd_hash = get_sd_hash(stream_info)
@@ -2451,7 +2476,11 @@ class Daemon(AuthJSONRPCServer):
         blob_hashes = [blob.blob_hash for blob in blobs]
         if need_sd_blob:
             # we don't want to use self._download_descriptor here because it would create a stream
-            sd_blob = yield self._download_blob(sd_hash, timeout=sd_timeout)
+            try:
+                sd_blob = yield self._download_blob(sd_hash, timeout=sd_timeout)
+            except Exception as err:
+                response = yield self._render_response(0.0)
+                defer.returnValue(response)
             decoded = read_sd_blob(sd_blob)
             blob_hashes = [blob.get("blob_hash") for blob in decoded['blobs']
                            if blob.get("blob_hash")]
@@ -2488,10 +2517,7 @@ class Daemon(AuthJSONRPCServer):
 def get_sd_hash(stream_info):
     if not stream_info:
         return None
-    try:
-        return stream_info['sources']['lbry_sd_hash']
-    except KeyError:
-        return stream_info.get('stream_hash')
+    return stream_info['sources']['lbry_sd_hash']
 
 
 class _DownloadNameHelper(object):
@@ -2510,7 +2536,7 @@ class _DownloadNameHelper(object):
     @defer.inlineCallbacks
     def setup_stream(self, stream_info):
         sd_hash = get_sd_hash(stream_info)
-        lbry_file = yield self.daemon._get_lbry_file_by_sd_hash(sd_hash)
+        lbry_file = yield self.daemon._get_lbry_file(FileID.SD_HASH, sd_hash)
         if self._does_lbry_file_exists(lbry_file):
             defer.returnValue(lbry_file)
         else:
@@ -2673,6 +2699,8 @@ class _GetFileHelper(object):
             return self.daemon._get_lbry_file_by_sd_hash(self.val)
         elif self.search_by == FileID.FILE_NAME:
             return self.daemon._get_lbry_file_by_file_name(self.val)
+        elif self.search_by == FileID.STREAM_HASH:
+            return self.daemon._get_lbry_file_by_stream_hash(self.val)
         raise Exception('{} is not a valid search operation'.format(self.search_by))
 
     def _get_json(self, lbry_file):
@@ -2806,7 +2834,7 @@ def report_bug_to_slack(message, installation_id, platform_name, app_version):
 
 
 def get_lbry_file_search_value(search_fields):
-    for searchtype in (FileID.SD_HASH, FileID.NAME, FileID.FILE_NAME):
+    for searchtype in (FileID.SD_HASH, FileID.NAME, FileID.FILE_NAME, FileID.STREAM_HASH):
         value = search_fields.get(searchtype)
         if value:
             return searchtype, value